refact(InputRadio): extract into its on module
