+++ apps/store/src/app/debugger/car-trial/CarTrialDebuggerForm.tsx
@@ -5,9 +5,9 @@ import { Text, Space } from 'ui'
 import { SubmitButton } from '@/appComponents/SubmitButton'
 import { ErrorMessages } from '@/components/FormErrors/ErrorMessages'
 import { InputCarRegistrationNumber } from '@/components/InputCarRegistrationNumber/InputCarRegistrationNumber'
+import * as InputRadio from '@/components/InputRadio/InputRadio'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
 import { PersonalNumberField } from '@/components/PersonalNumberField/PersonalNumberField'
-import * as InputRadio from '@/components/PriceCalculator/InputRadio'
 import { Field } from '@/features/carDealership/DebuggerCarTrial/debuggerCarTrial.types'
 import { createCarTrial } from './actions'
