+++ apps/store/src/app/debugger/trial/components/TrialContractForm.tsx
@@ -5,9 +5,9 @@ import { Space } from 'ui'
 import { Field } from '@/app/debugger/trial/debuggerTrial.types'
 import { SubmitButton } from '@/appComponents/SubmitButton'
 import { ErrorMessages } from '@/components/FormErrors/ErrorMessages'
+import * as InputRadio from '@/components/InputRadio/InputRadio'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
 import { PersonalNumberField } from '@/components/PersonalNumberField/PersonalNumberField'
-import * as InputRadio from '@/components/PriceCalculator/InputRadio'
 import { TextField } from '@/components/TextField/TextField'
 import { setupTrialContract } from '../actions'
 import { CoInsuredCount } from './CoInsuredCount'
