+++ apps/store/src/components/PriceCalculator/AutomaticField.tsx
@@ -1,3 +1,4 @@
+import * as InputRadio from '@/components/InputRadio/InputRadio'
 import { InputSelect } from '@/components/InputSelect/InputSelect'
 import { UseRegistrationAddressField } from '@/components/PriceCalculator/UseRegistrationAddressField'
 import { StepperInput } from '@/components/StepperInput/StepperInput'
@@ -9,7 +10,6 @@ import { CarMileageField } from './CarMileageField'
 import { CarRegistrationNumberField } from './CarRegistrationField'
 import { CurrentInsuranceField } from './CurrentInsuranceField/CurrentInsuranceField'
 import { ExtraBuildingsField } from './ExtraBuildingsField'
-import * as InputRadio from './InputRadio'
 import { PetCatBreedsField } from './PetCatBreedsField'
 import { PetDogBreedsField } from './PetDogBreedsField'
 import { useTranslateFieldLabel } from './useTranslateFieldLabel'
